Pattern database: corners for heuristics

+++ src/solve/move.go
@@ -33,17 +33,17 @@ func applyMove(c *cube.Rubik, move byte) {
 	cube.RotateFace(c, move_options[move], false)
 }
 
-func GetOppositeMove(move byte) int {
-	opposite := 0
-	optionsCountQuarter := (len(move_options) * 2) / 3
-	if int(move) < optionsCountQuarter {
-		if int(move) < optionsCountQuarter/2 {
-			opposite = int(move) + optionsCountQuarter/2
+func GetOppositeMove(move byte) byte {
+	opposite := byte(0)
+	optionsCountQuarter := byte((len(move_options) * 2) / 3)
+	if move < optionsCountQuarter {
+		if move < optionsCountQuarter/2 {
+			opposite = move + optionsCountQuarter/2
 		} else {
-			opposite = int(move) - optionsCountQuarter/2
+			opposite = move - optionsCountQuarter/2
 		}
 	} else {
-		opposite = int(move)
+		opposite = move
 	}
 	return opposite
 }
